show_room/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==4.0.4
django-bootstrap==0.2.4
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-registration==3.3
django-tinymce==3.4.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
whitenoise==6.1.0
